+++ imgaug/augmentables/polys.py
@@ -312,9 +312,6 @@ class Polygon(object):
         """
         return not self.is_out_of_image(image, fully=True, partly=False)
 
-    # FIXME this is not completely correct as it only evaluates corner points
-    #       and a polygon can still have parts of if inside the image, even when
-    #       all points are outside of it
     def is_out_of_image(self, image, fully=True, partly=False):
         """
         Estimate whether the polygon is partially or fully outside of the image area.
@@ -327,7 +324,7 @@ class Polygon(object):
             If a tuple, it is assumed to represent the image shape and must contain at least two integers.
 
         fully : bool, optional
-            Whether to return True if the polygon is fully outside fo the image area.
+            Whether to return True if the polygon is fully outside of the image area.
 
         partly : bool, optional
             Whether to return True if the polygon is at least partially outside fo the image area.
@@ -339,16 +336,12 @@ class Polygon(object):
             on defined parameters. False otherwise.
 
         """
+        # TODO this is inconsistent with line strings, which return a default
+        #      value in these cases
         if len(self.exterior) == 0:
             raise Exception("Cannot determine whether the polygon is inside the image, because it contains no points.")
-        inside = self._compute_inside_image_point_mask(image)
-        nb_inside = sum(inside)
-        if nb_inside == len(inside):
-            return False
-        elif nb_inside > 0:
-            return partly
-        else:
-            return fully
+        ls = self.to_line_string()
+        return ls.is_out_of_image(image, fully=fully, partly=partly)
 
     @ia.deprecated(alt_func="Polygon.clip_out_of_image()",
                    comment="clip_out_of_image() has the exactly same "